requirements.txt
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
jinja2==3.1.3
numpy==1.24.3 